Massive rewrite, for which I'm too tired to detail (yeah, I'll regret this five years from now ;).  In brief,
We now organize held postings by sender address and this "summary" is
what's presented when .../mailman/admindb/listname is visited.  The
pending subs and unsubs don't change, but the preamble is shorted
(with links to more detailed instructions).

For each group of held messages, you can click on one link to see the
details of the all the messages sent by a single address, or you can
click on a different link to view the details of just a single
message.  A third link lets you view the details of all the held
messages, just like the old admindb page used to give you.

The details page is just like the old admindb in form and function,
except that ADMINDB_PAGE_TEXT_LIMIT now works ;) and there are links
back to the summary page.

One additional neat hack is that on the summary page, there's an
option to add an email address to one of the auto sender filter lists
(viewable in the admin/privacy/sender screen).  This is only visible
if the address isn't already on one of the four filter lists.

Where this should be a boon is if you've got a bunch of messages being
held that are coming from the same address, and they all appear to be
spam.  You can discard them all in one fell swoop (no click-'n'-scroll
necessary!), and add them to the auto-discard list, so you never have
to worry about them again.

Note that if what you wanted to do was add a regexp to say the
auto-discard list, you'd have to first add this address, then go to
the admin/privacy/sender page and edit the address into a regexp
filter.  Easily done, while reducing the complexity of the admindb
summary page, which already pushes the edge of "too busy".