main(): Refactor calculation of QUERY_STRING values.
membership_options(): Refactor calculation of adminurl.  Also, add a
hack to hide or display the table legend (normally hidden) using
QUERY_STRING extensions.  This may not yet play nicely with all the
member chunking extensions.

Add Dan Mick idea to include an indication of the reason for
`nomail'.  Use an abbreviation, described in the legend.