Fix a race condition and a reference counting issue related to forced lock acquisition in mailmanctl.  Specifically,
_take_possession(): In the while-loop we need to not only check for
linkcount <> 2, but also that the contents of the lock file is our
magic token and not some other process's.  This fixes the race
condition.

_disown(): Set the private __owned flag to false so __del__ won't try
to finalize, and thus unlink the lock files.  This fixes the reference
counting problem in mailmanctl when both the failing non-forced lock
and the re-aquired force lock both point to the same files.