"import Cookie" instead of "from Cookie import SimpleCookie as Cookie"
The latter broke the upgrade script when upgrading from MM2.1a2 to
current cvs.  This is because bin/update imports Mailman.MailList
which imports Mailman.SecurityManager which, if /that/ tries to
from...import...as will fail with an import error since there's still
a residual old Cookie.py laying around that doesn't have a
SimpleCookie class.

Once bin/update does its thing and gets rid of Mailman/Cookie.py{,c}
we're safe, but for now this is the best way to worm around the
problem.