Some elaboration of the member moderation flag semantics.  Now it's possible to assign an action to the flag, which ought to work better for e.g. announce-only lists.  Specifically,
process(): If we're looking at a member and that member has their
moderation flag turned on, then we look at
mlist.member_moderation_action.   This is a 3-way selector specifying
whether the message should be held for approval, rejected (bounced),
or discarded.  There's also a member_moderation_notice attribute which
includes text to send back to the member when their posting is
automatically rejected (WIBNI we could include that text in any held
or discard notification?  Too much code disruption to do that right
now).