Several fixes to improve the bounce score testing when the administrator has adjusted the threshold (inspired by Dan Mick). Specifically,
ZEROHOUR_PLUSONEDAY: New constant that is supposed to represent the
day containing the first midnight after the epoch.  We'll add (0,)*6
to this tuple to get a value appropriate for time.mktime().

_BounceInfo.reset(): New method to encapsulate resetting stale bounce
scores.  Also used in the constructor.

registerBounce(): Implement refined rules for setting and checking the
bounce score.  As described in a message to mailman-developers:

    1) are we a member?  No, return
    2) do we have previous bounces?  No, register it, skip to step 7
    3) is the member disabled?  yes, return
    4) is the bounce info from today?  yes, don't increment, skip to step 7
    5) is the info stale?  yes, reset it and skip to step 7
    6) increment score for today's bounce
    7) is score > threshold?  yes, disable 'em.

disableBouncingMember(): Refactor since cron/disabled can call this
method too now.

__sendAdminBounceNotice(): msg argument can now be a Message instance
or a string.  Wrap in a MIME* instance as appropriate.