Massive simplification to avoid making all the BerkeleyDB calls ourselves.  Given the incompatible state of BDB libraries on most modern Linux distros, and given the sorry state of affairs in Python's BDB wrappers, this is the only sensible solution.
Essentially, I've kept all the plain text (data/aliases and
data/virtual-mailman) update code, and then we just call postalias and
postmap on the files to get the .db counterparts regenerated.  This
should be foolproof since only Postfix itself needs to use consistent
libraries -- attempting to match Postfix and Python would be a
nightmare.

I don't believe we're opening ourselves up to any vulnerabilities by
using os.system() in this case.

Also: all functions which are not meant to be part of the public MTA
interface are now prefixed by a single underscore.

I'm too tired to give detailed changes...