Someone -- I forgot who :( -- made a suggestion that list_members should be able to filter on nomail (i.e. delivery disabled) as well as the kind of digest being received.
I liked the idea, but wasn't keen on the new command line switches, so
I reimplemented it using optional arguments on the switches.  Sadly,
this means we can't use getopt, but still, it's not that hard to hand
code.  getopt-sig to the rescue!

Specifically, -d/--digest grow an optional argument, and added the
-n/--nomail switch.

statusp(): Returns a boolean specifying whether the nomail filter
matches the delivery status of the member.

main(): New option parsing code, and added the filters described
above.