GetConfigInfo(): Update the description of the interpolation strings to not include the %()s chrome.  We might be using the new experimental $-strings.  This leaves the instructions with a something left to be desired, but it should be okay for now, and I don't want to add two descriptions, one for %-strings and another for $-strings.
HandleForm(): We break the handling of the Autoresponse attributes
into a special method which does some sanity checking on the
interpolation strings.  First, it checks to see if we're using the new
experimental $-strings, and then depending on the style of
interpolation being used, it extracts the identifiers (variables) in
the string and compares them against a known valid list.  Any invalid
variables are flagged and the changes are discarded.  If the user
simply left off a trailing `s' (as in %(foo)s), then this can be
corrected, but a warning is still printed.