WHITESPACE -> whitespace everywhere.
ValidateEmail(): Add Dan Mick's suggestion to throw out email
addresses with embedded spaces in the (although the calculation of
such is done a bit differently).  This may not be a complete solution
because any address that's first sent through email.Utils.parseaddr()
will have embedded spaces collapsed anyway by rfc822.py's similarly
named function.  That latter is a bug that needs to be fixed in
Python, but that's not a high priority right now.  Also note that
"this is an email address"@dom.ain is (probably) legal and is handled
correctly by parseaddr() but will still be rejected by
ValidateEmail().  I think that's fine, as anybody who's got embedded
spaces in their address is probably going to have a miserable life
anyway.

to_dollar(), to_percent(), dollar_identifiers(),
percent_identifiers(): New functions to support experimental $-string
substitutions, and conversions between %-strings and $-strings.