add_error_message(): Removed as a duplicate of Document.addError(). Change all calls of add_error_message() to doc.addError().
main(): Fix the error reporting when no email address is given.  Also,
rework the error reporting for when the login page's unsub or remind
buttons are used.  Specifically, when rosters are public, we should
provide error messages when the given address is not a member, but
when rosters are private, lie about the confirmation message or
password reminder.