Patches to support duplicate suppression for explicit recipients, written by Ben Gertzfield, ported to MM2.1 by Marc MERLIN.
(Note: this patch was rewritten by Barry; parts of it was originally
in GUIBase, but that wasn't the right place for it.)

Specifically,

GetConfigInfo(): Add support for the new_member_options attribute,
which is a checkbox that the list owner can use to select the initial
values for the member options fields.

_setValue(): Handle setting the new_member_options attribute by
converting it from a list of option names to a bitfield.