_load(), _save(): Now that we're saving instances in the pending database, we need to use pickle instead of marshal.  Pickle's more reliable and portable anyway.
We can do auto-upgrade from marshal to pickle.  If the .pck file isn't
found, we'll read the .db file.  But we always write the .pck file and
if that's successful we can delete the .db file.  We also do the
standard "write a .tmp file and rotate with os.rename()" dance.