Make it much cheaper to call Load(), especially when the state hasn't changed since the last Load().  This means it will be cost effective to reload the state when necessary in qrunners which don't lock the list (e.g. OutgoingRunner).  Specifically,
InitTempVars(): Set a temporary timestamp attribute, which gets the
mtime of the config.pck file upon successful load.

__save(): When we save the file, set the timestamp to the file's mtime
(we just saved it so it must be up-to-date -- since this is done with
the list lock acquired, there shouldn't be a race).

__load(): If the file's mtime is <= the current timestamp, then we've
got the most current state.  This method can now return (None, None)
meaning we didn't need to load anything.

Load(): Watch for dict is None and e is None, meaning we didn't need
to load anything.