An adaptation of Marc MERLIN's fixes to actually honor SMTP_MAX_SESSIONS_PER_CONNECTION.  Specifically,
sendmail(): We can't return directly from the try-clause because then
the decrement of __numsessions won't be executed.  Instead, stash the
results in a local variable, twiddle numsessions, and then return the
results.

process(): When doing bulkdeliver, be sure to only decorate the
message once.  This prevents multiple headers and footers if the
message is requeued because of some temporary delivery problem.

Note: we don't need to do this for verpdeliver because each message is
crafted uniquely anyway.

Also, get rid of an unused local variable.