_postValidate(): New method which is called after all the form attributes have been processed.  This allows for the gui component to do sanity checks across inter-dependent values.  Default is no-op.
handleForm(): We don't need the `tag' arguments to addError() since
the default is 'Error: '.  Also, rearrange the exception logic so that
the _setValue() call happens in an else and the except clauses don't
need to return.  This allows the gui component to still get a chance
at _postValidate(), even if there were errors.