Ron Jarrell noted that the tagger logic was broken when messages start with indented text.  Fixes,
process(): Fix short-circuit test for mlist.topics_enabled instead of
mlist.topics.

scanbody(): We don't need to do much more than filter out the blank
lines when creating a mini-message to look for Subject: and Keywords:
fields in the body.  We /do/ need to create a subclassed
_ForgivingParser which simply stops processing instead of throwing
exceptions when it hits a line that's neither a header nor a
continuation.  WIBNI if we didn't have to cut-n-paste the bulk fo the
_parseheaders() method?