New architecture for email commands.  Instead of the monolithic (and unmaintainable) MailCommandHandler.py file, we've now got a framework where each command is implemented in a separate file.  This means it's both more extensible and more flexible:
- you can easily add new commands for things I haven't thought of
  <wink>, and the `help' command will automatically adjust

- you can disable commands entirely by removing the appropriate file

- you can disable, change, or add commands on a per-list (or even
  per-message or per-sender) basis

CommandRunner.py is the module that calls into this framework.  Each
command is implemented as a cmd_<command>.py file.  The `set' command
is the most complicated.  The help text is currently implemented as
module docstrings (for most commands), so the i18n catalogs must be
updated.  Also the help.txt files will be updated.