Support the new email commands framework.  We now no longer use MailCommandHandler.py.  Specific changes include:
class Results: Basically a bag to hold state during the processing of
a particular email message loaded with commands.  This makes it easy
to pass data around (like the mailing list object and the message).
It also contains the results of the command processing and gloms up
the response email.

_dispose(): This method actually gets simplified because it just needs
to pass things to the Results instance and let it do the job of
processing the email commands.