do_list_categories(): Teach this about the attribute transformations on the way out of the mailing list via getValue().  Always defer to getValue() before getattr() is used.  This lets us simplify the logic a bit, but we still take care to print useful representations of the data.  Use splitlines() instead of assuming lines are \n separated.
getPropertyMap(): New function to return the reverse mapping of
attribute names to the gui components that define them.

FakeDoc: Dummy class definition which defines an API used by the gui
components when an error occurs.

do_input(): Teach this about the attribute transformations on the way
into a mailing list via _setValue().  Yes, this is a non-public
method, but there seems to be no other way. :(