InitVars(): Add a new attribute hold_and_cmd_autoresponses which is a dictionary mapping sender addresses (they don't have to be members) to a 2-tuple of information (date, count).  date is a 3-tuple of (year, month, day) and count is the number of auto-responses sent to the sender today.
autorespondToSender(): This method is called whenever we're about to
send a response to a -request message, or a hold notification.  It
returns 1 if the caller should send a normal autoresponse, and 0
otherwise.  When we've just exceeded the maximum number of
autoresponses, we send a message indicating that no more responses
will be sent today (instead of the normal autoresponse).