InviteNewMember(): Prepend any text given in the mass subscribe page to the invitation notification.  Also, set an `invitation' attribute on the UserDesc record stored in the pending database.  This is a hack so that ProcessConfirmation() can figure out if the subscription request was an invitation or not (easier than adding an INVITATION type to the Pending module).
ApprovedAddMember(): Option text argument for passing in prepended
text for subscription notifications.

ProcessConfirmation(): If the SUBSCRIPTION action was an invitation
(based on the magic existance of the `invitation' attribute on the
UserDesc record), don't do additional approvals.  If the user was
invited, we /know/ the admin had to approve of that!