Fix a nasty bug that could cause disabled members to be deleted from the list too early!  This scenario:
- a member starts bouncing and gets an initial bounce score, but isn't
  yet disabled

- their mail starts working again for a while

- we start getting bounces for them again, but now their old info is
  stale, so we reset it

The bug was that when we reset the info, we actually didn't reset
their bounce score, we reset their "notifications-left" value!  That
means that if they accumulated more bounce scores and got disabled,
they'd immediate get removed from the list because they'd have no
bounce notifications left!  Ouch.

The fix is to enshrine the reset semantics in default arguments for
_BounceInfo.reset(): default for score is 0, default for date is None
(meaning use Day 1), default for notices left is None, is basically a
placeholder.

registerBounce(): If we're going to reset stale bounce info, use the
defaults, except for noticesleft, which gets its value from
bounce_you_are_disabled_warnings.

Thanks to Danny Terweij for being the guinea pig. ;)