Patch set for SF bug #596565.  Use symbolic user/group names instead of numeric ids.  Initial idea and patch by Todd Vierling, fleshed out by Barry.
Specific changes here:

check_caller(): The second argument now is a const char* instead of a
GID_T, since it takes a group name not an id.

Also, significantly rewrote the error messages that get logged when a
group mismatch (or other error) occurs.  Hopefully now when errors do
occur, they'll be much easier to understand because they'll actually
give the suggested configure options to fix the problem.