Better support for "funny" characters in subject prefixes.
encode_p(): Removed.

_isunicode(): Helper for unicode type testing.

prefix_subject(): Better algorithm for handling unicode subject
prefixes.  Now we always set the Subject header to a Header instance
if we're adding the prefix.  Always convert the prefix and all
previous Subject header bits to unicode (possibly with character
replacements) so that the Header class does the right thing, i.e. does
the us-ascii, charset hint, utf-8 encoding.  Because of this, if the
list's charset is us-ascii, we'll substitute iso-8859-1 for a slightly
wider character coverage.