__setstate__(): This plus email-2.4.3 should fix (most of) Peer Heinlein's recent problems.  This patch adds several things to better update pickled Message objects to the latest email package. Specifically,
__init__(): Add a version number to Message objects so that we can
make __setstate__() more efficient (I don't want to do the next bit of
hackery for every single unpickled Message object).

__setstate__(): For Message objects older than the current email
package version, we'll ensure that the data format is the same.  We
keep the previous updates and add one that trolls through the message
headers, ensuring that if any of the headers are Header instances,
that their chunks all have 2nd items as Charset instances, not
strings.  This is an invariant for email 2.4.3, but older Message
instances might not conform.