Martin v. Loewis's SF patch #634109 for better scrubbing of multipart messages.  Specifically,
process(): Each part that we're scrubbing, we'll set the content type
explicity to text/plain (since that's what it is now).  We also record
the character set of each part and if that's shared by all the
subparts, we'll just use that.  But, if we don't know the charset of
any of the parts then we'll use the list's preferred language's
charset.

Then we'll make sure all the text/plain parts have the same character
set, using 'replace' if necessary.  (We may eventually want to utf-8-ify
or html-entity-ify them in this case).

Patch slightly modified by Barry for i18n and style.