quick_maketext(): Copy the body of the try: from Utils.maketext() for completeness (this really needs to be refactored).
Article.as_html(): Set the encoding of the page to the list's
preferred encoding, which seems to make the headers and bodies come
out okay when mixing languages -- at least as in the minimal testing
I've been able to do.

Woo boy, I'm really not positive about this change and would love a
second opinion.  Martin?