setListIfUnset(): Hack to set self._mlist if it isn't already set. This is needed to fully fix SF bug #644294.  Old archived Articles won't have the _mlist attribute but we really want that for better handling of page creation.  Unfortunately, only the HyperDatabase knows this value so this is a public method it can call after the fact.
as_html(): Fix address obscuring in the article pages.  If
ARCHIVER_OBSCURES_EMAILADDRS is true, then we'll `at-ify' the visible
author text and point the mailto: to the list's posting address.

HyperArchive.__init__(): Pass the maillist to the HyperDatabase
constructor.

write_index_entry(): If ARCHIVER_OBSCURES_EMAILADDRS, `at-ify' the
author field in the index.  This might catch non-addresses with @'s in
them, but so what?  For the most part, it'll moderately obscure email
addresses.

__processbody_URLquote(): Futile attempt at code cleaning for a method
that I don't think ever gets called. :/