subscription_prompt(): Move the setting of the title to after the point where the i18n and doc languages are set.  Also, get the languages using GetLanguageDescr() so that the language pull down list is also in the correct language.
Fixes SF #658213 reported and patches by Daniel Buchmann.

subscription_cancel(): Set the language to the user's preferred so
that the cancel message is given in the right language.

subscription_confirm(): Set the i18n and doc languages to the language
from the form so it also shows up in the correct language.