I18n fixes for the digester, based on Tokio Kikuchi's patch #668819. I've modified it though, so any brokenness is my fault.  Specifically,
KEEP: Removed in favor of Default.py.in's MIME_DIGEST_KEEP_HEADERS.

process(): Use Generator.flatten() instead of the deprecated __call__
syntax.

send_i18n_digest(): Make sure things like the Subject headers in the
toc are coerced to the language the digest is being sent in (i.e. the
list's preferred language).  Make sure the masthead, header and footer
attachments also get the right matching charset attribute.  Make sure
that the username extracted from the From header is also in the right
character set.  Use the Header object for proper wrapping of the
Subject lines in the toc, not Utils.wrap() which was broken and not
i18n aware.

Add a blank line between entries in the toc (might be controversial).

Send messages in the plain text digest through the scrubber so
attachments (and their MIME goo) don't mess up the plain text digests.

lheader(): New convenience function.