Fixes for the cross-site scripting bug
http://online.securityfocus.com/archive/1/308154

Closes SF bug # 674533 by Tokio Kikuchi

Specifically,

main(), loginpage(): Check the `user' cgi var for validity and print
an innocuous (and non-privacy leaking) message if it fails that test.

Don't pass the entire cgidata object to loginpage; instead give just
the language argument which is all the latter function uses.

Also, be sure to use `safeuser' everywhere we print a message to the
results page.  safeuser is the escaped version of the `user' cgi var.

Unrelated:

main(): Sanity check the `language' cgi variable and use the mailing
lists's preferred language if it is deliberately invalid.