Tokio Kikuchi's SF patch #674401 to fix some issues with i18n in the digests.  Specifically,
lheader() -> oneline(): Also, get rid of **kws since we're not going
to use these.  Catch any LookupError or UnicodeError that may occur,
but do the .encode() with 'replace' for a better chance to succeed.

send_i18n_digests(): Use Utils.wrap() to wrap the subject headers, and
also in the kept headers in plain text digests.  Eliminate the blank
line between subjects in the ToC.