Use True/False where appropriate (so much cleaner!)
_dispose(): Add another key to the msg metadata, deliver_after.  If a
temporary failure occurs, this key will hold the time in epoch-seconds
after which we can retry delivery.  This stops us from retrying every
second as reported in SF bug # 699900.  Default is to retry once per
hour (see Defaults.py.in).

Also, use copy.deepcopy() instead of converting the message to a
string and re-parsing it.