DEAL_WITH_PERMFAILURES_EVERY: Do this only every 10 times through the loop.
_dispose(): Remove the logic for retrying temporary failures from here
to the RetryRunner.  This eliminates total cpu consumption on the
re-queuing of messages with tempfailures (as can happen with delivery
via Postfix to non-existent local addresses in its default
configuration).

Restructure the SomeRecipientsFailed clause and removed the
deliver_after calculation -- we don't need it.  DELIVERY_RETRY_WAIT is
also removed.

_cleanup(): When shutting down the qrunner, be sure to register all
deferred permanent failures.

Backport candidate.