Move list_lists and list_members over to Mailman.bin package and rewrite to use $-strings and optparse.
A side effect of the rewrite is that list_members -d and -n now require their
argument, whereas before they were optional (but that is counter to command
line parsing standards and not supported by optparse).  Now, instead of bare
-n use '-n any' and instead of bare -d use '-d any'.

Fix ReopenableFileHandler.emit() to ensure that all lines end in a newline.

In i18n.py, create and use a subclass of string.Template so that dotted
attribute paths are legal in $-strings.  So for example, if you have an object
'foo' in one of the namespaces and that has an attribute 'bar', you can say
"$foo.bar" to print the value of foo.bar.  Should reduce the number of
extraneous local variables just to do the attribute path traversal.