Fix test_message.py by finishing the wind-through of the configuration object and fixing the invocation and shutdown of mailmanctl.  While the tests in this module work individually, they do not yet work as a group.
-C added to testall.py, and mailmanctl now passes that flag on to qrunner.

UserNotification sets reduced_list_header in the msgdata, but the behavior of
this flag has changed.  It used to suppress List-Help, List-Subscribe, and
List-Unsubscribe as well as List-Post and List-Archive.  However, List-Help,
List-Subscribe and List-Unsubscribe should definitely be included in
UserNotifications, and List-Post has a different variable controlling it now.
Therefore, always add List-Help, List-Subscribe, and List-Unsubscribe.

Some style updates to Message.py