Fix some buglets with virtual domain support and repair unit tests broken by this change.  More unit tests should be added.
misc/sitelist.cfg is removed -- this is an ex-site list.

MailList.GetNoReplyEmail() -> MailList.no_reply_address (property)

UserNotification._enqueue(), OwnerNotification._enqueue(): when queing the
message to the virgin queue, be sure to use the fully qualified (i.e. posting)
address for the list.

In the MTA modules, be sure to set up the target of the mail commands as the
fqdn listname because otherwise we can't find the correct list.  This needs
some tweaking/testing for Postfix's virtual domain support.

MailList.Load() has to grow an optional argument specifying the fqdn
listname.  The problem is that in some situations, we can't calculate that
because we don't know _internal_name, so it has to be passed in.  This is
mostly the case in the MailList ctor where a Load hasn't happened yet.  For
backward compatibility though, if it's not passed in, just use
mlist.fqdn_listname.