Fix several problems reported by Mark.  First, Python 2.3 doesn't have built-in sets and sadly we still have to support it.  Utils.list_names() therefore now uses a sets.Set if it finds the built-in set type missing.
main() in list_lists.py also has to list()-ify the set in order to sort it.

There's a bootstrapping problem when bin/update tries to re-situate a
non-qualified list name to a qualified list name.  The problem is that the
MailList ctor will call Load() with the nonqual name, which in turn will call
CheckVersion() by default.  The latter will then turn around and try to call
Load() with the fqdn list name, but that list name won't yet exist since it
hasn't been situated.

The solution (but maybe not the best one) then is to pass the check_version
flag to the MailList ctor, which gets passed down to the Load() call.  Only
situate_list() will call this to avoid the version checking before the list
has been situated.