First crack at an XML exporter of a mailing list's configuration and membership.  The next step is to write an XML importer that reads this file. I'm not 100% sure that all the import data is included yet, but the intent is that this will be the official way to move mailing lists.
A few notes: member passwords are not included by default, the idea being that
if we enable XML dumping from the web, we don't want the clear text user
passwords to be leaked.  A command line option includes the member passwords.

Also, the various substitutable texts (i.e. those that include %-strings) will
be autoconverted to $-strings.  In Mailman 2.2, we'll only have $-strings,
although this is not yet enforced in other parts of the code yet.

Convert config_list.py to mmshell, $-strings, and optparse.